requirements.txt
amqp==5.1.1
anyjson==0.3.3
asgiref==3.6.0
astroid==2.15.2
asttokens==2.2.1
attrs==22.2.0
Automat==22.10.0
autopep8==2.0.2
Babel==2.12.1
backcall==0.2.0
beautify==0.1.4
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
constantly==15.1.0
cryptography==40.0.1
cssselect==1.2.0
decorator==5.1.1
dill==0.3.6
Django==4.2
django-admin-sortable2==2.1.5
django-cache-machine==1.2.0
django-celery==3.1.17
django-cors-headers==3.14.0
django-debug-toolbar==4.0.0
django-extensions==3.2.1
django-filter==23.1
django-js-asset==2.0.0
django-mptt==0.14.0
django-redis==5.2.0
django-redis-cache==3.0.1
django-uuslug==2.0.0
djangorestframework==3.14.0
djangorestframework-datatables==0.7.0
djangorestframework-simplejwt==5.2.2
executing==1.2.0
filelock==3.11.0
flake8==6.0.0
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.4
ijson==3.2.0.post0
incremental==22.10.0
ipython==8.12.0
isort==5.12.0
itemadapter==0.8.0
itemloaders==1.0.6
jedi==0.17.2
jmespath==1.0.1
kombu==5.2.4
lazy-object-proxy==1.9.0
lxml==4.9.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mypy==1.2.0
mypy-extensions==1.0.0
numpy==1.24.2
packaging==23.0
pandas==2.0.0
parsel==1.7.0
parso==0.7.1
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.2.0
pluggy==1.0.0
prompt-toolkit==3.0.38
Protego==0.2.1
psycopg2-binary==2.9.6
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
PyDispatcher==2.0.7
pydocstyle==6.3.0
pyflakes==3.0.1
Pygments==2.14.0
PyJWT==2.6.0
pylama==8.4.1
pylama-pylint==3.1.1
pylint==2.17.2
pylint-django==2.5.3
pylint-plugin-utils==0.7
pyOpenSSL==23.1.1
python-dateutil==2.8.2
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
python-memcached==1.59
python-slugify==8.0.1
pytz==2023.3
queuelib==1.6.2
redis==3.5.3
requests==2.28.2
requests-file==1.5.1
Scrapy==2.8.0
scrapy-djangoitem==1.1.1
service-identity==21.1.0
six==1.16.0
snowballstemmer==2.2.0
sorl-thumbnail==12.9.0
sqlparse==0.4.3
stack-data==0.6.2
text-unidecode==1.3
tldextract==3.4.0
tomli==2.0.1
tomlkit==0.11.7
traitlets==5.9.0
Twisted==22.10.0
twisted-iocpsupport==1.0.3
typing_extensions==4.5.0
tzdata==2023.3
ujson==5.7.0
urllib3==1.26.15
vine==5.0.0
w3lib==2.1.1
wcwidth==0.2.6
wrapt==1.15.0
zope.interface==6.0
